Parse raw_data on error only if the Content-Type is application/json.
Some opensearch REST APIs will return a Content-Type of text/plain for certain
status codes which will trigger logging when the response is not parsable as
json. Checking the content-type first avoids this error.

This frequently happens with 401s, which occurs when we rotate user keys.
The input that triggers this behavior can be triggered with `curl`:

```
$  curl -k -u"admin:wrong" -w"\n%{http_code}\n%{content_type}\n" https://localhost:9200/_cluster/health
Unauthorized
401
text/plain; charset=UTF-8
```

Signed-off-by: Trevor McCulloch <trevor.mcculloch@benchling.com>